Watch for positions to change
Map would not update until something else triggered a draw like zooming or dragging. Set up a watcher on the positions prop to force a redraw when it updates. Also added a flag to ignore the first time the watcher fires since it seemed like it was firing before everything was initialized