Minor drawing optimizations, remove loading images
Minor optimization for the drawing call to only calcualte bounding boxes when zooming in
Removes the canvas image source prop and opts to pass in an Image directly. No more loading anything directly in this component. Everything should be passed into this component fully loaded
removed the deep watcher for positions as it seemed to fire more than necessary, and the ignoring of the first watcher since everything should be loaded before the canvas component is called